MAGECLOUD-3357: [Cloud Docker] Changes in .magento.app.yaml should affect php modules
