+++ src/Docker/Compose/FunctionalCompose.php
@@ -126,6 +126,6 @@ class FunctionalCompose extends ProductionCompose
      */
     protected function getPhpExtensions(string $phpVersion): array
     {
-        return self::DEFAULT_PHP_EXTENSIONS;
+        return PhpExtension::DEFAULT_PHP_EXTENSIONS;
     }
 }
